+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.13.0] - 2025-11-13
+
 ### Fixed
 
 - country labels in search country input for `TelephoneNumberInput` includes also national prefix.
@@ -634,7 +636,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/3.12.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/3.13.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -659,6 +661,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[3.13.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.12.0...3.13.0
 [3.12.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.11.2...3.12.0
 [3.11.2]: https://github.com/neolution-ch/react-hook-form-components/compare/3.11.1...3.11.2
 [3.11.1]: https://github.com/neolution-ch/react-hook-form-components/compare/3.11.0...3.11.1